Python/APP1/Python_Bootcamp.Day_05/src/requirements.txt
python3
flask
PyYaml
atexit